simplified logic and tests
